+++ CHANGELOG.md
@@ -1563,3 +1563,4 @@
 - [2026-05-28 20:42] feat(analytics): integrate custom telemetry logger in analytics
 - [2026-05-28 21:46] test(theme): implement widget integration tests for theme
 - [2026-05-28 22:25] feat(forms): integrate custom telemetry logger in forms
+- [2026-05-28 22:27] feat(navigation): integrate custom telemetry logger in navigation